Fix time series transformer
